add styles and move files around
